[refactor](common): topAppBar -> common/topappbar 패키지로 이동 (#23)
