+++ app/src/main/java/com/kuit/ourmenu/ui/addmenu/screen/AddMenuTagScreen.kt
@@ -39,7 +39,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.kuit.ourmenu.R
-import com.kuit.ourmenu.ui.addmenu.component.AddMenuTopAppBar
+import com.kuit.ourmenu.ui.common.topappbar.OurMenuBackButtonTopAppBar
 import com.kuit.ourmenu.ui.addmenu.component.IconSelectBottomSheet
 import com.kuit.ourmenu.ui.addmenu.component.TagSelectBottomSheet
 import com.kuit.ourmenu.ui.common.CustomTextField
@@ -65,7 +65,7 @@ fun AddMenuTagScreen(modifier: Modifier = Modifier) {
     BottomSheetScaffold(
         scaffoldState = scaffoldState,
         topBar = {
-            AddMenuTopAppBar {
+            OurMenuBackButtonTopAppBar {
                 Text(
                     text = stringResource(R.string.add_menu),
                     style = ourMenuTypography().pretendard_600_18,
