+++ app/src/main/java/com/kuit/ourmenu/ui/home/screen/HomeScreen.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.kuit.ourmenu.ui.home.component.recommendation.sub.HomeSubRecommendation
-import com.kuit.ourmenu.ui.common.OurMenuLogoTopAppBarWithAddButton
+import com.kuit.ourmenu.ui.common.topappbar.OurMenuAddButtonTopAppBar
 import com.kuit.ourmenu.ui.home.component.recommendation.main.HomeMainRecommendation
 import com.kuit.ourmenu.ui.home.dummy.HomeDummyData
 
@@ -24,7 +24,7 @@ fun HomeScreen(modifier: Modifier = Modifier) {
 
     Scaffold(
         topBar = {
-            OurMenuLogoTopAppBarWithAddButton()
+            OurMenuAddButtonTopAppBar()
         }
     ) { innerPadding ->
         Column(
